split activities lib in sub modules to simplify functions dependency graph
